Don't include development dependencies in the image
Currently the `Dockerfile` used to build the container image installs
the development dependencies (go and git, mostly) in the image itself,
and then builds the binary using them. This patch changes the build
process so that the binary is built outside the `Dockerfile`, and then
just copied to the image. To achive this the `Dockerfile` is now
generated from a `Dockerfile.in` template, using simple variable
substitution in the new `image` target in the `Makefile`. With this
approach the size of the image goes from 650 MiB to 289 MiB.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>